revert commit 1791
That commit broke bsp/stm3210x. The cause is that we calls rtgui_system_init in rtgui_startup. But the calibration starts before rtgui_startup and rtgui_system_init. This lead to crash.

One way to deal with it is move rtgui_system_init out of rtgui_startup and call it manually. But it will affect other bsps. So just revert the "bad" commit.

git-svn-id: https://rt-thread.googlecode.com/svn/trunk@1793 bbd45198-f89e-11dd-88c7-29a3b14d5316