[msh] cd to path when possible
When DFS_USING_WORKDIR is enabled, msh will try to chdir if the command
is not found. For example, if there is folder named "SD", and there is
no `SD` command and `SD.mo` module, typing "SD" in msh will cd to the
folder.

In more detail, the execution order right now is:

1. built-in command
2. module
3. cd to path

If you have a folder named "ps", you have to run `cd ps` to change the
working dir.