msh can run now
the key step is to swap the order of rt_interrupt_nest -- and RT_OBJECT_HOOK_CALL(rt_interrupt_leave_hook,()) in irq.c. This is an improvised solution and an issue has been raised.

rt_memset is also changed to adapt to both 8bit 16bit addressing. Not fully tested.