AArch64: extract Kconfig with libcpu
Now all the libcpus' Kconfig is config in libcpu/Kconfig. This will make the
configuration difficult for an arch, you know, there is so many errata and
extensions' codes will config in ARMv8. Well, DMA and memory even PCI's
configuration is different from arch. So it's not a good idea to shove these
features in libcpu/Kconfig.

Signed-off-by: GuEe-GUI <GuEe-GUI@github.com>