[smart] replace page management algorithm in varea
Previous implementation of page recycling in varea slow down the
searching of free page especially with the large varea region. This
patch use a page table entry query instead to improve the performance
on varea reconstructions.

Also fixup: dfs mmap

The return value of DFS mmap is not guarantee to be identical of the
kernel mmap address, since the varea can changed during the mmap.
So this patch use the address as the return value directly as the
return value.

Signed-off-by: shell <smokewood@qq.com>