[smart] fixups of lwp recycling
feat: cleanup recycling routine of lwp.
The tty is not recycled on exit() currenttly. This patch add a new
layer of resource recycling including address space, fd tables, etc.
Also, the pid is allocated from parent proc, so it's much reasonable to
recycle the pid in parent process. So it's moved to waitpid() now.

fixup: trap fixup in normal context

Signed-off-by: Shell <smokewood@qq.com>