[bsp][nxp/mcxn] Add support for MCXN CTIMER PWM driver.
This patch adds support for PWM generation using CTIMER on MCXN devices.

Each CTIMER provides 4 PWM channels, channels sharing the same
CTIMER instance will have the same period settings, and the duty
cycle will be maintained if period is changed through one of the four
channels.

The period channel is automatically assigned and will be transferred
to one of the other available channels if that channel is used for PWM
generation, no glitches should be expected during the transition.

The patch also provides a sample PWM output configuration connected to
the on-board green LED which can be enabled through Kconfig.

Signed-off-by: Yilin Sun <imi415@imi.moe>