bsp: qemu-virt64-riscv: Kconfig: remove duplicated RT_USING_VIRTIO
There is problem in the Kconfig of qemu-virt64-riscv.
RT_USING_VIRTIO is already defined in components/drivers/virtio/Kconfig,
so it should not be defined again in the BSP kconfig.

The solution is to rename RT_USING_VIRTIO to BSP_USING_VIRTIO.
Also, refer to the configuration method for virtio device drivers in
bsp/qemu-virt64-aarch64/drivers/Kconfig to rewrite the configuration
of BSP_USING_VIRTIO_xxx.

Updated default .config/rtconfig.h at the same time.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>