bsp: cvitek: optimze build operations for aarch64
For bsp/cvitek, duo256m, the operation process is
slightly different when building images for riscv
and arm64.

To improve the user experience, unify the building
process and steps of the two as follows:

- Enter c906_little and execute `scons`

- Enter cv18xx_riscv/cv18xx_aarch64 and execute `scons`

That's all. Finally, we can get `fip.bin` and
`boot.sd` under `bsp/cvitek/output/milkv-duo256m`.

Update the README.md accordingly.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>