doxygen: reorganize directory structure
The original directory structure was too flat and
inconvenient to view.

The directory is now reorganized according to the
first-level chapters, and some sub-chapters are
moved to the second-level directory.

Note: documentation/env is not moved to
documentation/4.tool because there are too many
external places that reference this path,
especially in bsp, so this directory path remains
as it is.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>