bsp: k230: use configuration instead of immediate value
The original code used immediate values directly when it
came to memory layout, which was not a good programming
habit.

Now we use macros and try to be compatible with SDK
configuration values in the SDK compilation environment.

The main changes are to clean up board.h, and extract
some memory layout constants to define a new header
file mem_layout.h.

Also update KERNEL_VADDR_START to the default value for
Smart, so we can use it as base to calculate other mapping
address.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>