doxygen: cleanup and re-org files
This patch contains two small changes, mainly to
prepare for the next step of organizing the modules
framework.

The first change: move the files related to doxygen
examples to `documentation/0.doxygen/example/`

The second change: delete `documentation/0.doxygen/mainpage.h`,
which is the mainpage of the old page. Now the new mainpage
is `documentation/INDEX.md`, so the duplicate can be
deleted.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>