bsp: k230: use rttpkgtool to package kernel
The original method of packaging and burning kernel
depends on installing the K230 RTOS SDK. It is quite
cumbersome. Now the relevant logic is extracted and
a small tool rttpkgtool is created to package the
rtthread.bin.

Considering that rttpkgtool still needs to compile
opensbi when packaging k230, which means that a
separate cross-toolchain needs to be installed.
Therefore, in order not to affect the CI operation,
the current packaging logic is not integrated into
POST_ACTION in rtconfig.py, but requires kernel
developers to run the build script separately for
packaging.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>