MAINTAINER: standardize tag writing
Update the tag writing requirements in the Note.
It is recommended that the tag name directly refer
to the module path in the code tree. For example,
if the module path is 'bsp/cvitek', then the tag
name should be 'bsp_cvitek'. And the name should
be lowercase.

According to the above rules, the tag name of the
existing entries are adjusted. Other changes remain
unchanged.

Other text modifications are mainly to limit the
width of each line, try not to exceed 80 columns.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>