doxygen: update doc for env to latest
The introduction to Env under the documentation
section of the RTT source code repository is
outdated and unmaintained. We recommend moving the Env
documentation to https://github.com/RT-Thread/env/
for self-maintenance.

Solution:
Direct users to the latest README at https://github.com/RT-Thread/env/.

The Env content under the documentation section of RTT
will not be deleted for now, but will be marked as "outdated".

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>