utest: fixed UTEST_DEVICE_TC
In 01e015450f, "utest: move driver-related case under to drivers,"
UTEST_DEVICE_TC was replaced with RT_UTEST_DRIVERS_CORE. However,
some files are forgot to do this replacement, so this patch address
this.

Note: In .config, lines like "# CONFIG_XXX is not set" do not need
to be modified; they are comments. Furthermore, for BSPs, running
"scons --menuconfig" and save the config will automatically replace
them. Therefore, no modification is necessary in this case.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>