[libcpu-riscv]: [common64 virt64]: Add the dual-core feature under the SMP architecture.
Currently, it does not support operation with two or more cores.

Solution: Since the system runs in S-mode and does not support access to the
mhartid register, the hartid is currently stored in the satp register (this
register is not used when the bsp qemu-virt64-riscv runs RT-Thread). Additionally,
logic for storing boot_hartid and multi-core initialization logic for the sp
pointer have been added in startup_gcc.S. Logic for secondary core wake-up and
entry has been added in cpuport.c.

Signed-off-by: Mengchen Teng <teng_mengchen@163.com>