[libcpu-riscv]: [surpport SMP]: Fix issues
Fix some potential concurrency issues and code standardization problems.
The .percpu section is only used when both ARCH_MM_MMU and RT_USING_SMP
are enabled. However, there is a certain amount of space waste since no
macro guard is added for it in the link script currently.

The physical memory of QEMU started in CI is 128MB, so RT_HW_PAGE_END is
modified from the original +256MB to +128MB.

Modify the SConscript file under the common64 directory to include
common/atomic_riscv.c in the compilation process.

Signed-off-by: Mengchen Teng <teng_mengchen@163.com>